requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
joblib>=1.3
pyyaml>=6.0
matplotlib>=3.7

# Optional — for FastAPI inference server
fastapi>=0.110
uvicorn>=0.29
pydantic>=2.0

# Optional — for database / S3 ingestion
# sqlalchemy>=2.0
# s3fs>=2024.1
